refactor : file dir changed for export library (and change tsconfig , package.json)
