+++ packages/@marulloc-shopify-next-api/package.json
@@ -3,6 +3,12 @@
   "version": "1.0.0",
   "private": true,
   "exports" : { 
+    "./v24.01": "./src/v24.01/index.ts",
+    "./v24.01/@shopify-graphql": "./src/v24.01/@shopify-graphql/index.ts",
+    "./v24.01/@shopify-types": "./src/v24.01/@shopify-types/index.ts",
+    "./v24.01/@toolkit-types": "./src/v24.01/@toolkit-types/index.ts",
+    "./v24.01/services": "./src/v24.01/services/index.ts",
+    "./v24.01/utils": "./src/v24.01/utils/index.ts"
   },
   "dependencies": { 
     "typescript": "^5",
