+++ packages/@marulloc-shopify-next-api/src/v24.01/@shopify-graphql/fragments/collection.ts
