+++ packages/@marulloc-shopify-next-api/src/v24.01/@shopify-types/shopify-search.ts
