+++ packages/@marulloc-shopify-next-api/src/v24.01/@toolkit-types/toolkit-collection.ts
@@ -1,4 +1,4 @@
-import { ShopifyCollection } from '../../@shopify-types/shopify-collection';
+import { ShopifyCollection } from '../@shopify-types/shopify-collection';
 
 export type ToolkitCollection = ShopifyCollection & {
   handleRoute: string;
