+++ packages/@marulloc-shopify-next-api/src/v24.01/@toolkit-types/toolkit-page.ts
@@ -1,0 +1,3 @@
+import { ShopifyPage } from '../@shopify-types/shopify-page';
+
+export type ToolkitPage = ShopifyPage;
