+++ packages/@marulloc-shopify-next-api/src/v24.01/@toolkit-types/toolkit-product.ts
@@ -1,5 +1,5 @@
-import { ShopifyImage } from '../../@shopify-types/shopify-common';
-import { ShopifyProduct, ShopifyProductVariant } from '../../@shopify-types/shopify-product';
+import { ShopifyImage } from '../@shopify-types/shopify-common';
+import { ShopifyProduct, ShopifyProductVariant } from '../@shopify-types/shopify-product';
 
 export type ToolkitProduct = Omit<ShopifyProduct, 'variants' | 'images'> & {
   variants: ShopifyProductVariant[];
