+++ packages/@marulloc-shopify-next-api/src/v24.01/@toolkit-types/toolkit-shop.ts
@@ -1,4 +1,4 @@
-import { ShopifyLocalization, ShopifyMenuItem, ShopifyShopInfo } from '../../@shopify-types/shopify-shop';
+import { ShopifyLocalization, ShopifyMenuItem, ShopifyShopInfo } from '../@shopify-types/shopify-shop';
 
 export type ToolkitMenu = Array<ShopifyMenuItem>;
