+++ packages/@marulloc-shopify-next-api/src/v24.01/services/product/parser.ts
@@ -1,7 +1,7 @@
 import { ShopifyProduct } from '../../@shopify-types/shopify-product';
 import { flatConnection } from '../../utils/flat';
 import { generateGIDRoute, generateHandleRoute } from '../../utils/routeGenerator';
-import { ToolkitProduct } from '../@toolkit-types/toolkit-product';
+import { ToolkitProduct } from '../../@toolkit-types/toolkit-product';
 
 export const parseProduct = (product: ShopifyProduct): ToolkitProduct => {
   const { images, variants, ...rest } = product;
