+++ packages/@marulloc-shopify-next-api/src/v24.01/services/search/service.ts
@@ -2,9 +2,9 @@ import { getPredictiveSearchQuery, getSearchResultQuery } from '../../@shopify-g
 import { ShopifyLocaleContext } from '../../@shopify-types/shopify-common';
 import { ShopifyProductFilter } from '../../@shopify-types/shopify-search';
 import { flatConnection } from '../../utils/flat';
-import { storeFetch } from '../../utils/storeFetch';
-import { ToolkitProduct } from '../@toolkit-types/toolkit-product';
-import { ToolkitPredictiveSearch, ToolkitSortKey } from '../@toolkit-types/toolkit-search';
+import { storeFetch } from '../../storeFetch';
+import { ToolkitProduct } from '../../@toolkit-types/toolkit-product';
+import { ToolkitPredictiveSearch, ToolkitSortKey } from '../../@toolkit-types/toolkit-search';
 import { parseProducts } from '../product/parser';
 import { parsePredictiveSearch, parseSortParams } from './parser';
 import { GetPredictiveSearchService, GetProductsSearchService } from './types';
