+++ packages/@marulloc-shopify-next-api/src/v24.01/services/search/types.ts
