+++ packages/@marulloc-shopify-next-api/src/v24.01/utils/routeGenerator.ts
