--- packages/@marulloc-shopify-next-api/v24.01/index.ts
@@ -1,1 +1,0 @@
-
