--- packages/@marulloc-shopify-next-api/v24.01/services/@toolkit-types/toolkit-page.ts
@@ -1,3 +1,0 @@
-import { ShopifyPage } from '../../@shopify-types/shopify-page';
-
-export type ToolkitPage = ShopifyPage;
